meta-qti-automotive-distro: add debug flags for easier debugging
The debug flags ensure that binaries are built with debug symbols.
The debug flags also provides several compilation flags that make GDB
easier to use on the target.

DEBUG_FLAGS does not change the optimization level used to compile code
with gcc. Additionally, the debug symbols are stripped from the binaries
and placed in separate -dbg packages, so there should be no size or
performance regression to the main binaries.

DEBUG_FLAGS adds the following compilation flags as of Yocto 5.0:
 1. -g - compiler produces debugging information. Yocto strips
    debugging information and places it in -dbg package.

 2. -fmacro-prefix-map=<build machine path>=<target path> -
    compiler fixes code macros __FILE__ and __BASE_FILE__ as if the
    files resided at the target path rather than the build machine
    path. Ensures logs do not print build machine paths.

 3. -fdebug-prefix-map=<build machine path>=<target path> -
    compiler saves the target paths in the produced debug symbols
    rather than saving build machine path. Ensures GDB knows how to
    find the source code on the target that was used to create the
    binary being debugged.

 4. -feliminate-unused-debug-types - compiler eliminates debug symbols
    for types not used in the code. No impact on the binaries used
    in production.

Change-Id: Id8d1e53de6ec0d416517a075e1fe6942600d17c9
Signed-off-by: Eric Smith <ericsmit@qti.qualcomm.com>